Move composition layer from EssentialFeed to EssentialApp
