+++ EssentialApp/EssentialApp/Comment/CommentUIComposer.swift
@@ -9,6 +9,7 @@
 import Foundation
 import UIKit
 import EssentialFeed
+import EssentialFeediOS
 
 public final class CommentUIComposer {
 	private init() {}
