+++ EssentialFeed/EssentialFeediOS/Comment UI/Controllers/CommentViewController.swift
@@ -9,40 +9,13 @@
 import UIKit
 import EssentialFeed
 
-
-final class WeakRefVirtualProxy<T: AnyObject> {
-	private weak var object: T?
-	
-	init(_ object: T) {
-		self.object = object
-	}
-}
-
-extension WeakRefVirtualProxy: CommentLoadingView where T: CommentLoadingView {
-	func display(_ viewModel: CommentLoadingViewModel) {
-		object?.display(viewModel)
-	}
-}
-
-extension WeakRefVirtualProxy: CommentView where T: CommentView {
-	func display(_ viewModel: CommentViewModel) {
-		object?.display(viewModel)
-	}
-}
-
-extension WeakRefVirtualProxy: CommentErrorView where T: CommentErrorView {
-	func display(_ viewModel: CommentErrorViewModel) {
-		object?.display(viewModel)
-	}
-}
-
 public protocol CommentViewControllerDelegate {
 	func didRequestCommentRefresh()
 	func didCancelCommentRequest()
 }
 
 public final class CommentViewController: UITableViewController, CommentErrorView, CommentLoadingView {
-	var delegate: CommentViewControllerDelegate?
+	public var delegate: CommentViewControllerDelegate?
 	
 	private var tableModel = [CommentCellController]() {
 		didSet {
